Enhanced player view; added team
